Separate userinfo.js from route.js. PROMISE WORKING
