+++ public/index.html
@@ -27,9 +27,9 @@
      <div class="user-container" ng-repeat="user in ctrl.users">
        <img class="avatar" src="{{ user.avatar_url }}">
        <div id="attribute-container">
-         <a class="username" ng-href="{{ user.html_url}}"  target='_blank'> {{ user.login }} </a>
-         <p class="repos"> Repos: {{ user.public_repos}} </p>
-         <p class="followers"> Followers: {{ user.number_followers}} </p>
+         <a class="username" ng-href="{{ user.html_url }}"  target='_blank'> {{ user.login }} </a>
+         <p class="repos"> Repos: {{ user.public_repos }} </p>
+         <p class="followers"> Followers: {{ user.number_followers }} </p>
        </div>
      </div>
    </div>
