+++ public/js/services/SearchService.js
@@ -3,12 +3,11 @@ userApp.service('SearchService', ['$http', function($http) {
   this.getUser = function(username) {
 
     return $http({
-      url: "/api/userinfo",
+      url: "/user",
       method: 'GET',
       params: {user_name: username}
       })
       .then(function(response) {
-        console.log(response.data);
         return response.data;
       })
       ;
